Add size requirements for collections
There are three requirements added: one for lists, one for sets, and one
for collections. The  only difference between them, and the reason for
their existence (as in fact the collection requirement can be used for
lists and sets), resides in the fact that the custom error message
explicitly mentions the collection type.