Add nonzero requirements
There are two: on for chars, and one for numbers.
The latter is typed wide: Number, to accept any numerical boxed
primitive in one go (instead of needing one per numeric primitive type).
This can be done because the error message is identical no matter the
type, so it can all be factored into one requirement.

There is no requirement regarding the boolean type, as restricting it
would make no sense.