Configure the Sonatype publishing plugin
The server ID is explicitly set to the default value, because explicit
is better than implicit (not my words!)
Only the required checksums are generated.
The deployment and package names are explicitly set to a value that
contains the project's ID and version to better find one's way around
deployments on Sonatype's dashboard page.