Change the internals of the field class
The atomic reference member is replaced with two volatile fields: a
value holder and a boolean flag indicating whether the field has been
set. This allows to abstract from the idea that value == null => unset.
Since an atomic reference is no longer used, synchronization has to be
performed explicitly. An altered version of the double-checked locking
pattern is used: the if blocks test the flag field, and the blocks are
not nested (rather, the bodies of the if blocks exit the method).
Other methods are modified to use the flag field to check if the value
is set, as intended.