Do not apply the requirements if the value is null
There is no functional sense in having requirements applicable to null.
It could even be a cause for exceptions, if requirements are not
implemented in a null-safe manner; but I feel that I owe to users to
alleviate their task, and skip the null-hardening, because it is so easy
to do.
The internal method checking the value before setting it is reorganized
so that the this-check and the requirement-applying loop are only
performed if the value is not null. Also, a comment is added
highlighting the fact that no requirements are applicable to null
values.