Update tests for method hashCode
The tests no longer check that the stable field's hash code equal that
of its value, as the method no longer declares to do so. Instead, they
only check that the method return a non-zero value, either if the value
is set or unset. (Hardly more can be tested, now that the method no
longer specifically documents what it returns)