Add tests to check parallel runs of set/trySet
The tests run each setter in parallel, attempting to ensure that the
setter behaves soundly  when called this way (a sort of light stress
test).

This method yields inherently unpredictable results, so it may or may
not result in a full coverage of the method (and apparently more often
not than does).

There is however a workaround to make the test work more systematically,
which I want to mention to keep a formal record of it: by wrapping the
Thread.start calls in a synchronized block, locking on the stable
field's internal lock (and by sleeping for a short time, e.g. 100 ms,
giving time for the two threads to reach the synchronized block in the
setter). This requires making the internal lock visible, which I do not
plan to do (as this would defeat the purpose of using an internal lock
in the first place).

Te implement the workaround, simply make StableField.lock public and, in
tests methods testSetCalledInParallelSucceedsOnce and
testTrySetCalledInParallelSucceedsOnce, replace:

    thread1.start();
    thread2.start();

with:

    synchronized (sf.lock) {
    	thread1.start();
    	thread2.start();
    	try {
    		Thread.sleep(100);
    	} catch (IntgerruptedException e) {
    		fail(e);
    	}
    }

The call to sleep leaves time for the two threads to perform the initial
checks in the respective setter and reach the synchronized block, having
made sure that they both found the set flag set to false (this is
crucial!) waiting to lock on the lock monitor that the test method
currently holds.

This makes sure that one of the two threads will fail the second check
of the set flag.