Remove the "delayed" setter methods
These methods could not be made to work in a satisfying manner with the
concurrency safety guarantee. I was faced with a technical dilemma,
unable to reconcile the requisite that the supplier would only be called
to actually set the field's with the current, thread-safe implementation
of the setters (double-check locking pattern). Either the supplier was
called before the synchronized block, thus risking to be invoked only
to find that another thread had initialized the field, or it was called
within the synchronized block, introducing all the risks brought about
by running untrusted code while holding an exclusive lock. And not to
mention where to place the call to the internal check method among this.

Moreover, the more I thought about these methods, the less I could
picture a legitimate use case for them. Maybe when the acquisition of
the value is costly, but then, if one can afford the cost of this
operation to initialize the field, one can afford it too to *fail* to
initialize. This could be solved by introducing an isSet method, but I
will not do this! This class is not an Optional clone! Failing to
initialize the field is supposed to be an exceptional condition, not an
ordinary situation. Maybe trySet should go as well; at least it's
compatible with the double-check lock.

I guess I only introduced these methods in an esthetic impulse, to
balance the introduction of getOpt and map (functional getters /
functional setters). However it was unfounded, as I hope to have shown.