Define setters/getters for the volatile members
The purpose of this evolution is to make the original class StableField
unsynchronized--as it originally was--so that users who do not require
synchronization have the option of avoiding the overhead. A subclass
will be added later, and this one will perform synchronization. This
way, synchronization or not will be an explicit choice of users.
These methods will be the way that the subclass performs
synchronization, hence the protected visibility.