Refactor test instance creation in StableFieldTest
A non-test method is added to the test class to create an instance of
the class under test, instead of calling the constructor in the test
method.
This does not apply to constructor tests, as in these methods it is the
constructor itself that is tested, and the call should then not be
refactored away.