Override test cases for the shared field subclass
The override is defined simply by extending the original test cases, and
overriding the methods creating the test instances, returning a shared
instance rather than a base one. This is great! The only downside is
that there are requirements tests that do not use a stable field
instance, and are thus duplicated needlessly. This is a negative
consequence of the design of the test case: test requirement instances
through a stable field rather than independently--these are more
integration tests than true unit tests. However, this is a minor issue
that I can bear to live with (although I may rewrite these tests in true