Access the members volatilely only in shared class
The volatile modifier is removed from the member declaration for set and
value in the base field class, so that accesses from this class do not
suffer the propagation overhead.
The members in the shared field subclass are no longer accessed
directly, but through their respective var handles, using methods that
perform the operation (get or set) in a volatile manner, so that the
concurrence-safety is preserved.