Consider member set in method equals
The presence of the member in the equality expression actually makes a
difference only in the specific case described in the previous commit
and made possible by the ALLOW_NULL requirements: when the two value
references are both null, but when one has been initialized and not the
other.
This case is now legitimate (because of the Requirements instance) and
therefore ought to be covered.